chore: safe bot removal.

+++ packages/bot/botkit/src/bot-factory/fs-bot-snapshot-storage.ts
@@ -71,7 +71,9 @@ export class FSBotSnapshotStorage implements BotSnapshotStorage {
 
   async deleteBackupedBot (id: string): Promise<void> {
     const botStorage = join(this._dir, id);
-    await fs.promises.unlink(botStorage);
+    if (fs.existsSync(botStorage)) {
+      await fs.promises.unlink(botStorage);
+    }
   }
 
   async reset (): Promise<void> {
